backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
websockets==13.1
python-multipart==0.0.12
requests>=2.31.0
urllib3>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
